Add nodelet implementation. (#41)
* Add nodelet implementation.

Add additional nodelet layer to mirror the node
implementation. This allows the Slam GMapping
library to be run as a nodelet instead. This
would allow you to, for example, run it under
the same nodelet manager as the nodelet producing
the /scan output for greater efficiency.

* Remove superfluous semicolons

Removed superfluous semicolons and
mildly clarified info stream output.